# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fleet visibility, session history, and control plane for AI agent fleets",
        "title": "Flightdeck Query API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/v1/access-tokens": {
            "get": {
                "description": "Returns every access_tokens row. Hash and salt are never exposed; the raw token itself is available only on the POST response at creation time (D095).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "access-tokens"
                ],
                "summary": "List access tokens",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.AccessTokenRow"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Mints an opaque ftd_ bearer token. The plaintext token is returned exactly once in the response; the platform cannot recover it afterwards. See DECISIONS.md D095.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "access-tokens"
                ],
                "summary": "Create an access token",
                "parameters": [
                    {
                        "description": "Token name",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateAccessTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/store.CreatedAccessTokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/access-tokens/{id}": {
            "delete": {
                "description": "Deletes the access_tokens row. The seeded Development Token row is protected and returns 403. Sessions previously authenticated with the token keep their token_name for auditability (sessions.token_id is set to NULL via ON DELETE SET NULL). See DECISIONS.md D095.",
                "tags": [
                    "access-tokens"
                ],
                "summary": "Revoke an access token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates the name field on the access_tokens row. Historical sessions keep their original token_name snapshot. The Development Token row is protected and returns 403.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "access-tokens"
                ],
                "summary": "Rename an access token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New name",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.RenameAccessTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.AccessTokenRow"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/admin/reconcile-agents": {
            "post": {
                "description": "One-shot operation that makes the agents table correct: recomputes total_sessions, total_tokens, first_seen_at, and last_seen_at against sessions ground truth (per-agent transaction; partial errors surface in the response errors array), then deletes orphan rows whose post-reconcile total_sessions = 0 AND last_seen_at \u003c NOW() - orphan_threshold_secs. Pass orphan_threshold_secs=0 to skip the delete step (counters-only). Default threshold is 30 days. Values \u003c 60 s rejected with 400 to prevent reaping freshly-upserted agents that the worker has not yet wired up to a session_start.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Reconcile agent rollup counters AND reap stale orphan rows",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Override the 30d default for the orphan-delete step. 0 skips deletion (counters-only). Values 1..59 rejected with 400.",
                        "name": "orphan_threshold_secs",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reconcile completed; errors array is empty",
                        "schema": {
                            "$ref": "#/definitions/store.ReconcileResult"
                        }
                    },
                    "207": {
                        "description": "Reconcile completed with per-agent errors (see errors array)",
                        "schema": {
                            "$ref": "#/definitions/store.ReconcileResult"
                        }
                    },
                    "400": {
                        "description": "orphan_threshold_secs malformed or absurdly small",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid bearer token",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Token is valid but lacks admin scope",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Another reconcile is already in progress",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Fatal database error (agent listing or pool exhaustion)",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/agents": {
            "get": {
                "description": "Returns agents matching the supplied filters. Multi-value filters (state, agent_type, client_type, hostname, user, os, orchestration) accept comma-separated values and repeated query params; values within a dimension are OR, values across dimensions are AND. ``search`` is a case-insensitive substring match against ``agent_name`` and ``hostname``. ``updated_since`` filters on ``last_seen_at \u003e= ts``. State is computed via LATERAL subquery against the most-recent session. Pagination defaults to 25/page, max 100.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "agents"
                ],
                "summary": "List agents with filters, search, sort, and pagination",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by agent_type (repeatable/comma: coding, production)",
                        "name": "agent_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by client_type (repeatable/comma: claude_code, flightdeck_sensor)",
                        "name": "client_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by rollup state (repeatable/comma: active, idle, stale, closed, lost)",
                        "name": "state",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by hostname (repeatable/comma, exact match)",
                        "name": "hostname",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by user_name (repeatable/comma, exact match)",
                        "name": "user",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by sessions.context.os (EXISTS subquery; matches if ANY session had the OS)",
                        "name": "os",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by sessions.context.orchestration (EXISTS subquery)",
                        "name": "orchestration",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Case-insensitive substring search on agent_name + hostname",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to agents whose last_seen_at \u003e= this ISO-8601 timestamp",
                        "name": "updated_since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort column (default: last_seen_at). One of: last_seen_at, first_seen_at, agent_name, total_sessions, total_tokens, state, user, hostname",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "asc or desc (default: desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max results (default 25, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset for pagination (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.AgentListResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid filter value, sort column, order direction, or limit over 100",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid bearer token",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/analytics": {
            "get": {
                "description": "Flexible GROUP BY analytics endpoint. Returns time series data grouped by dimension.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analytics"
                ],
                "summary": "Query analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Metric: tokens, sessions, latency_avg, latency_p50, latency_p95, policy_events, estimated_cost (default: tokens)",
                        "name": "metric",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Dimension: flavor, model, framework, host, agent_type, team, provider (default: flavor)",
                        "name": "group_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Time range: today, 7d, 30d, 90d, custom (default: 30d)",
                        "name": "range",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Start time ISO 8601 (required when range=custom)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "End time ISO 8601 (required when range=custom)",
                        "name": "to",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Granularity: hour, day, week (default: day)",
                        "name": "granularity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to specific flavor",
                        "name": "filter_flavor",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to specific model",
                        "name": "filter_model",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to specific agent_type",
                        "name": "filter_agent_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to specific provider (anthropic, openai, google, xai, mistral, meta, other)",
                        "name": "filter_provider",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.AnalyticsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/directives": {
            "post": {
                "description": "Issues a shutdown or custom directive to a single agent session or all sessions of a flavor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "directives"
                ],
                "summary": "Create directive",
                "parameters": [
                    {
                        "description": "Directive definition",
                        "name": "directive",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.DirectiveRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/store.Directive"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.CreateAccessTokenRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.RenameAccessTokenRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "store.AccessTokenRow": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_used_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prefix": {
                    "type": "string"
                }
            }
        },
        "store.AgentListResponse": {
            "type": "object",
            "properties": {
                "agents": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/store.AgentSummary"
                    }
                },
                "has_more": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "store.CreatedAccessTokenResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "prefix": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    }
}